feat: allow stencil starter to be downloaded from on prem registry
When working behind a corporate proxy, requests to github.com might be rejected by the proxy. Often companies who use a corporate proxy also have an on-prem registry for npm packages. This PR allows to set the host URL to get the starter app from which allows stencil to be initialized behind a strict proxy.