chore(npm): add stub testing script
add a 'stub' testing script to `package.json`. the create-stencil cli
currently has a release process that differs from stencil core in that
it (in theory) uses `np` for performing releases (suggested by the
'release' script in `package.json`). however, 'release' will not run
prior to this commit due to a lack of 'test' command. this suggests to
me the package was version bumped and released manually.

to push the project to a uniform (and arguably safer) publish process,
add a stub testing script so that 'release' (which is a proxy for `np`)
does not fail. the project does not have any tests at the moment.
although this is tech-debt that must begin to be rememdied, we don't add
any in this commit for the sake of unblocking publish