chore(github): add ionitron to the repo
this commit adds our internal bot 'ionitron' to the create-stencil
repository. tehcnically, ionitron is already configured for the repo,
but needs a configuration file (this commit) to run.

this commit adds the ionitron configuration from the stencil core
compiler repository at
https://github.com/ionic-team/stencil/blob/9bfef1ad4637fe82c349eadbd9153e04417b1337/.github/ionic-issue-bot.yml.
it is a direct copy to keep the behavior similar between repositories
(making the behavior easier to grok for maintainers)