refactor(interactive): update indentation of getChoices()
this commit makes a small update to the code (rather than the output) of
getting the different starter choices. this is a relatively arbitrary
change, but one i made before i made further modifications to the
function to keep the two separate while i was working.

normally, i'd have a formatter like prettier handle this for us. since
it's not a part of the project at this time, i updated the function to
make it (subjectively) more readable until we get a single source of
truth in the project (to be done in STENCIL-447)