feat(starters): designate community packages
this commit adds the ability for the starter selection ux to denote
which starter projects are community driven.

this is accomplished by adding an optional `isCommunity` field to the
pre-existing starter metadata, and setting that field to `true` for the
application starter.

starters that are designated as community-driven will appear a such in
the picker:

```
? Select a starter project. Starters marked as [Community] are developed by the Stencil Community, rather than Ionic. For more information on the Stencil Community, please see https://github.com/stencil-community › - Use arrow-keys. Return to submit.

❯  component          Collection of web components that can be used anywhere
   [Community] app    Minimal starter for building a Stencil app or website
```

the disclaimer/explanation of the community nomenclature must be placed
on one line at this time. if any newlines appear in this string, the
explanation will reprint on every arrow key movement