refactor(prompts): install prompts package (#69)
this commit adds the prompts package to the project. previously, this
package had been vendored in 7177cc94d05cf07e6eac2981ff12c761a654face

while beginning to work on the codebase, it became clear we were working
with a ~3.5 year old version of the library. rather than keep up to date
with the library and merge pieces in as a needed, we go the route of
letting npm manage this for us. this will come at the cost of a larger
bundle size. however, given that this is a cli application, we believe
the cost to be worth the benefit