feat(community): be able to designate starters as community projects (#68)
this commit adds the ability for the starter selection ux to denote
which starter projects are community driven. this will help users
determine which starter projects are maintained by the Stencil team/
Ionic and those maintained by the community.

this is accomplished by adding an optional `isCommunity` field to the
pre-existing starter metadata, and setting that field to `true` for the
application starter.