chore(ci): add ci to the project
this commit adds an initial continuous integration (ci) pipeline to the
project that utilizes github actions. it is designed to only run on
ubuntu, using node 16 (the latest node runtime available) at this time,
as we don't believe there is a benefit to adding a "strategy" matrix at
this time when performing a build.

ci will be invoked on every pull request to verify a changelist adheres
to the checks the ci workflow contains today (just building the app), as
well as any check we add in the future (such as formatting, tests, etc.)
it will also run when a change is merged into the main branch.

this pipeline only builds the application at this time