chore(repo): add formatting infrastructure
this commit adds the official ionic preffier configuration to the
project. as a result, we delete the pre-existing .prettierrc.json file,
in order to use the configuration that's found in the
`@ionic/prettier-config` package.

prettier is added as an exact dependency of the project, as there may be
differences in the way prettier formats a codebase dependent on which
minor version is installed.

two npm scripts have been added in this commit - one for performing a
'dry-run' of the formatter, where divergences from the formatter's rules
are detected but not written to disk (useful for continuous integration
checks). another script has been written to run the formatter and write
changes to disk for each file that violates its rules