chore(test): add testing to project
add testing infrasturcture to the project. ts-jest was picked due to its
dominance in the ts ecosystem to tranpile typescript to javascript
(over, say babel). this decision was slightly arbitrary - no design doc
or case study was performed in the making of this decision.

we start with ts-jest's default configuration options, which we can tune
as we go.

this commit exercises the new infrastructure by getting `starters.ts`
under test