chore(dependencies): upgrade to typescript 4.7 (#74)
this commit upgrade the project to use typescript 4.7. any new
flags that were added to tsconfig.json's schema between v4.0
and v4.7.2 (inclusive) have been added on the 'strict' side of
things (here we prefer stricter settings in an effort to make the
code as "correct" as possible.

alphabetized the tsconfig.json file here. on the fence if
it provides actual value. my (personal) heurisitic is to alphabetize
these types of configs unless there's a some value in grouping
specific fields and placing them out of order (e.g. to make it clear
that a group of configurations together achieve a specific event).