fix(cli): fix --info flag
this commit restores the --info flag, which prior to this commit would
crash.

this commit updates the `getPkgVersion` helper to throw a slightly more
helpful error when either the package.json file cannot be found or (for
some reason) the 'version' field isn't found in the file. we don't
expect the latter to ever happen, but since reading a file returns
`any`, better safe than sorry. as a part of this update, the file
containing this function was moved to typescript.

to get `version.ts` under test, a new `getPackageJson` abstraction was
added. because the CLI can be run in dev mode (where
`version.ts#getPkgVersion` sits a directory below `package.json) and in
prod mode (where the transpiled `getPkgVersion` in `index.js` sits
adjacent to `package.json`), mocking getting the `package.json` file is
difficult when trying to exercise both the success and failure cases.
adding this abstraction allows us to mock the new function that wraps
the `require` statement much easier.

as a result of adding the abstraction, jest began to fail trying to read
`fs.rename` when importing `util.ts` in the `version.spec.ts` file.
because we're simply promisifying `fs.rename`, remove it in favor of the
native `fs.promises.rename`, which has been available since node 10