chore(volta): add volta to the project
this commit adds volta to the project to specify the versions of node
and npm that individuals working on the project should be using. this
does not affect the `package.json#engines` field nor runtime version
requirements of the package.

node 16 was chosen to align with what we run against in our continuous
integration (ci) pipeline. at this time, our ci pipeline runs in github
actions, which do not support node 18 at this time