chore(dev-deps): update typescript to v4.8.4
this commit updates typescript to v4.8.4, to align with upcoming changes
in the main repository where stencil core is built with v4.8 as well.

this commit also updates the dependabot manifest to ignore versions
greater than or equal to v4.9, as it is not supported in the stencil
ecosystem (at this time).