chore(links): update slack links
this commit updates the link to the stencil slack in the ionitron
response for support requests. the heroku app is no longer
working/maintained, and in it's place we link folks directly to the
slack instance.

this commit also replaces the ionic worldwide link (where that slack
instance is no longer used) with the global discord one