feat(init): initialize a git repo on creation
this commit updates the workflow for creating a new project to check
for/create a new git repository. specifically, it:
1. ensures that the user has git on their path
2. checks that the user is not in an existing git tree
    a. a new tree is not created if one is found
3. intializes a new git repo
4. adds/commits all generated files

the motivation for this pull request is to set up git repo's on behalf
of users, eliminating a step in the creation process. this change has an
additional benefit for the stencil team by allowing minimal reproduction
cases created with this cli to separate the boilerplate of a project
from the reproduction related changes (via git commits).

git/prettier settings were updated to ignore the generated `coverage/`
directory from jest as well