chore(ci): add automated dev releases
this commit implements an initial workflow for dev releases for the cli.
it is capable of:
- generating a build based using the same logic as other ci workflows
- generating a dev build version
- invoking a publish action with the dev version & 'dev' tag

a new publish action is added that is capable of incrementing the
version of the package and publishing changes to npm has been added. the
intent is to reuse this in some capacity for production releases