feat(starter): allow on-prem github instances (#358)
split the generation of the github url to pull starter templates from
into smaller, more precisely defined functions. 

add support for self hosted github repositories. when generating the url
to download a starter from, check for the appropriate environment
variable first. if it exists and it truthy, use that value